+++ packages/backend/src/models/PinnedListModel.ts
@@ -1,10 +1,10 @@
 import {
-    CreateChartPinnedItem,
-    CreateDashboardPinnedItem,
-    DeleteChartPinnedItem,
-    DeleteDashboardPinnedItem,
+    CreatePinnedItem,
+    DeletePinnedItem,
     isCreateChartPinnedItem,
+    isCreateSpacePinnedItem,
     isDeleteChartPinnedItem,
+    isDeleteSpacePinnedItem,
     NotFoundError,
     PinnedItem,
     PinnedList,
@@ -12,14 +12,18 @@ import {
 } from '@lightdash/common';
 import { Knex } from 'knex';
 import {
-    DbPinnedChart,
-    DbPinnedDashboard,
+    DbPinnedItem,
     DbPinnedList,
     PinnedChartTableName,
     PinnedDashboardTableName,
     PinnedListTableName,
+    PinnedSpaceTableName,
 } from '../database/entities/pinnedList';
-import { isDbPinnedChart } from '../utils';
+import {
+    isDbPinnedChart,
+    isDbPinnedDashboard,
+    isDbPinnedSpace,
+} from '../utils';
 
 type PinnedListModelDependencies = {
     database: Knex;
@@ -51,15 +55,18 @@ export class PinnedListModel {
         };
     }
 
-    async addItem(
-        item: CreateChartPinnedItem | CreateDashboardPinnedItem,
-    ): Promise<void> {
+    async addItem(item: CreatePinnedItem): Promise<void> {
         const results = await this.upsertPinnedList(item.projectUuid);
 
         if (isCreateChartPinnedItem(item)) {
             await this.database(PinnedChartTableName).insert({
                 pinned_list_uuid: results.pinnedListUuid,
                 saved_chart_uuid: item.savedChartUuid,
+            });
+        } else if (isCreateSpacePinnedItem(item)) {
+            await this.database(PinnedSpaceTableName).insert({
+                pinned_list_uuid: results.pinnedListUuid,
+                space_uuid: item.spaceUuid,
             });
         } else {
             await this.database(PinnedDashboardTableName).insert({
@@ -69,13 +76,16 @@ export class PinnedListModel {
         }
     }
 
-    async deleteItem(
-        item: DeleteChartPinnedItem | DeleteDashboardPinnedItem,
-    ): Promise<void> {
+    async deleteItem(item: DeletePinnedItem): Promise<void> {
         if (isDeleteChartPinnedItem(item)) {
             await this.database(PinnedChartTableName)
                 .delete()
                 .where('saved_chart_uuid', item.savedChartUuid)
+                .andWhere('pinned_list_uuid', item.pinnedListUuid);
+        } else if (isDeleteSpacePinnedItem(item)) {
+            await this.database(PinnedSpaceTableName)
+                .delete()
+                .where('space_uuid', item.spaceUuid)
                 .andWhere('pinned_list_uuid', item.pinnedListUuid);
         } else {
             await this.database(PinnedDashboardTableName)
@@ -92,18 +102,17 @@ export class PinnedListModel {
         };
     }
 
-    static convertPinnedItem(
-        data: DbPinnedChart | DbPinnedDashboard,
-    ): PinnedItem {
+    static convertPinnedItem(data: DbPinnedItem): PinnedItem {
         return {
             pinnedListUuid: data.pinned_list_uuid,
             pinnedItemUuid: data.pinned_item_uuid,
             savedChartUuid: isDbPinnedChart(data)
                 ? data.saved_chart_uuid
                 : undefined,
-            dashboardUuid: isDbPinnedChart(data)
-                ? undefined
-                : data.dashboard_uuid,
+            dashboardUuid: isDbPinnedDashboard(data)
+                ? data.dashboard_uuid
+                : undefined,
+            spaceUuid: isDbPinnedSpace(data) ? data.space_uuid : undefined,
             createdAt: data.created_at,
         };
     }
@@ -118,7 +127,6 @@ export class PinnedListModel {
             throw new NotFoundError('No pinned list found');
         }
 
-        const pinnedListUuid = [list][0].pinned_list_uuid;
         const pinnedCharts = await this.database(PinnedChartTableName)
             .select(
                 'pinned_list_uuid',
@@ -126,7 +134,7 @@ export class PinnedListModel {
                 'saved_chart_uuid',
                 'created_at',
             )
-            .where('pinned_list_uuid', pinnedListUuid);
+            .where('pinned_list_uuid', list.pinned_list_uuid);
         const pinnedDashboards = await this.database(PinnedDashboardTableName)
             .select(
                 'pinned_list_uuid',
@@ -134,12 +142,22 @@ export class PinnedListModel {
                 'dashboard_uuid',
                 'created_at',
             )
-            .where('pinned_list_uuid', pinnedListUuid);
+            .where('pinned_list_uuid', list.pinned_list_uuid);
+        const pinnedSpaces = await this.database(PinnedSpaceTableName)
+            .select(
+                'pinned_list_uuid',
+                'pinned_item_uuid',
+                'space_uuid',
+                'created_at',
+            )
+            .where('pinned_list_uuid', list.pinned_list_uuid);
 
         const pinnedList = PinnedListModel.convertPinnedList(list);
-        const pinnedItems = [...pinnedCharts, ...pinnedDashboards].map(
-            PinnedListModel.convertPinnedItem,
-        );
+        const pinnedItems = [
+            ...pinnedCharts,
+            ...pinnedDashboards,
+            ...pinnedSpaces,
+        ].map(PinnedListModel.convertPinnedItem);
 
         return { ...pinnedList, items: pinnedItems };
     }