+++ packages/frontend/src/components/SchedulerModals/SchedulerModalBase/SchedulerForm.tsx
@@ -1,9 +1,8 @@
 import {
     Button,
+    Classes,
     Colors,
     FormGroup,
-    HTMLSelect,
-    Icon,
     NumericInput,
     Radio,
     RadioGroup,
@@ -11,7 +10,6 @@ import {
     Tabs,
 } from '@blueprintjs/core';
 import { Tooltip2 } from '@blueprintjs/popover2';
-import cronstrue from 'cronstrue';
 import React, { FC, useEffect, useMemo, useState } from 'react';
 import useHealth from '../../../hooks/health/useHealth';
 import { useSlackChannels } from '../../../hooks/slack/useSlackChannels';
@@ -20,6 +18,7 @@ import { isInvalidCronExpression } from '../../../utils/fieldValidators';
 import { Limit, Values } from '../../DownloadCsvPopup';
 import { ArrayInput } from '../../ReactHookForm/ArrayInput';
 import AutoComplete from '../../ReactHookForm/AutoComplete';
+import CronInput from '../../ReactHookForm/CronInput';
 import Form from '../../ReactHookForm/Form';
 import Input from '../../ReactHookForm/Input';
 import Select from '../../ReactHookForm/Select';
@@ -190,16 +189,8 @@ const SchedulerSettings: FC<
     );
     const health = useHealth();
 
-    const cronValue = methods.watch('cron', '0 9 * * 1');
-    const cronHelperText = useMemo(() => {
-        const validationError =
-            isInvalidCronExpression('Cron expression')(cronValue);
-        const cronHumanString = cronstrue.toString(cronValue, {
-            verbose: true,
-            throwExceptionOnParseError: false,
-        });
-        return validationError ?? cronHumanString;
-    }, [cronValue]);
+    const isAddSlackDisabled = disabled || slackState !== SlackStates.SUCCESS;
+    const isAddEmailDisabled = disabled || !health.data?.hasEmailClient;
 
     return (
         <Form name="scheduler" methods={methods} {...rest}>
@@ -212,12 +203,9 @@ const SchedulerSettings: FC<
                     required: 'Required field',
                 }}
             />
-            <Input
-                label="Cron expression (UTC)"
+            <CronInput
                 name="cron"
-                placeholder="0 9 * * 1"
                 defaultValue="0 9 * * 1"
-                helperText={cronHelperText}
                 disabled={disabled}
                 rules={{
                     required: 'Required field',
@@ -304,13 +292,18 @@ const SchedulerSettings: FC<
                         >
                             <Button
                                 minimal
-                                onClick={() => append({ channel: '' })}
+                                className={
+                                    isAddSlackDisabled
+                                        ? Classes.DISABLED
+                                        : undefined
+                                }
+                                onClick={
+                                    isAddSlackDisabled
+                                        ? undefined
+                                        : () => append({ channel: '' })
+                                }
                                 icon={'plus'}
                                 text="Add slack"
-                                disabled={
-                                    disabled ||
-                                    slackState !== SlackStates.SUCCESS
-                                }
                             />
                         </Tooltip2>
                         <Tooltip2
@@ -334,11 +327,17 @@ const SchedulerSettings: FC<
                         >
                             <Button
                                 minimal
-                                onClick={() => append({ recipients: '' })}
+                                onClick={
+                                    isAddEmailDisabled
+                                        ? undefined
+                                        : () => append({ recipients: '' })
+                                }
                                 icon={'plus'}
                                 text="Add email"
-                                disabled={
-                                    disabled || !health.data?.hasEmailClient
+                                className={
+                                    isAddEmailDisabled
+                                        ? Classes.DISABLED
+                                        : undefined
                                 }
                             />
                         </Tooltip2>