+++ packages/frontend/src/components/common/ResourceList/ResourceTypeUtils.ts
@@ -1,22 +1,53 @@
 import {
     assertUnreachable,
     DashboardBasicDetails,
+    Space,
     SpaceQuery,
 } from '@lightdash/common';
 
 export enum ResourceListType {
     CHART = 'chart',
     DASHBOARD = 'dashboard',
+    SPACE = 'space',
 }
 
-type AcceptedResources = SpaceQuery | DashboardBasicDetails;
+type AcceptedResources = SpaceQuery | DashboardBasicDetails | Space;
+
+export type ResourceListChartItem = {
+    type: ResourceListType.CHART;
+    data: SpaceQuery;
+};
+
+export type ResourceListDashboardItem = {
+    type: ResourceListType.DASHBOARD;
+    data: DashboardBasicDetails;
+};
+
+export type ResourceListSpaceItem = {
+    type: ResourceListType.SPACE;
+    data: Space;
+};
 
 export type ResourceListItem =
-    | { type: ResourceListType.CHART; data: SpaceQuery }
-    | { type: ResourceListType.DASHBOARD; data: DashboardBasicDetails };
+    | ResourceListChartItem
+    | ResourceListDashboardItem
+    | ResourceListSpaceItem;
 
-export const wrapResource = (
-    resource: AcceptedResources,
+export type ResourceListItemCanBelongToSpace =
+    | ResourceListChartItem
+    | ResourceListDashboardItem;
+
+export const isResourceListItemCanBelongToSpace = (
+    resource: ResourceListItem,
+): resource is ResourceListItemCanBelongToSpace => {
+    return (
+        resource.type === ResourceListType.CHART ||
+        resource.type === ResourceListType.DASHBOARD
+    );
+};
+
+export const wrapResource = <T extends AcceptedResources>(
+    resource: T,
     type: ResourceListType,
 ): ResourceListItem => {
     switch (type) {
@@ -24,6 +55,8 @@ export const wrapResource = (
             return { type, data: resource as SpaceQuery };
         case ResourceListType.DASHBOARD:
             return { type, data: resource as DashboardBasicDetails };
+        case ResourceListType.SPACE:
+            return { type, data: resource as Space };
         default:
             return assertUnreachable(type, `Unknown resource type: ${type}`);
     }